optimizer: Rename flatmap_to_map.rs to flat_map_elimination.rs to match the transform therein
