+++ src/transform/src/canonicalization/flat_map_elimination.rs
